DTSPB-1184 - Refactored callback response transformer
